Added runTests(...) method  code and calls to XML_Stream_handler to validate XML written.
validated shopping list XML was written correctly.

Added and commented example code in Main activity in onCreate for shoppinglist and receipe.  These now persist to storage when setting name, instructions, and adding ingredient

defaulted isCrossedOut property in Ingredient to false and removed the setter in the instnatiation

found a bug in ingredient where countPartial was not gotten from the getter correctly - countfull was returned

Found and fixed a bug in the XML stream handler which caused the output stream to not write for recipes.