changed isFilePresent in datalayer for performance reasons, instead of getting all the files in the directory and looping through them to see if the iterated name matches the argument, instead simply attempt to open the file for input.  If a filenotfound exception is thrown, the file doesn't exist.
 modfied the email body text for recipe and shopping list to instead use discrete ingredient variables names rather than the tostring method

 created a createEmptyRecipe funciton which works the same as the createEmptyShoppingList.  it creates a recipe entry and autoincremenets as needed.  Added and commented the demonstration code for it in main activity in onCreate.

changed the empty constructor access for recipe and shopping list to private to avoid confusion at the UI layer.  The datalayer and business processes need empty constructors to operate, but I don't think the UI does.

modified the new shopping list activity to use the shoppinglist(this) constructor and added a comment there.