bugfixes,
modify ingredients to not have leading spaces in the display

taking a look at item deleted in shopping list crashes app.

Change the menu icon in new shopping list to the (lovely!) shopping cart +

Also added a little code to newshoppinglist and recipelist to save the name of the list or recipe when focus on the title edit text is lost.